Crate portfolio_management

The portfolio_management program provides accounts to profit on a bond based trading algorithm operating between different chains. The program is intended to be run fully on Solana (on-chain).

Documentation

Current documentation is limited to the main instructions related to bond creation logics and token bridge implementation

See the docs for instructions for token bridge implementation which uses wormhole.

See the docs for portfolio_management for portfolio related (bond creation and management) implementations.

Re-exports

pub use context::*;

pub use state::*;

pub use message::*;

pub use constants::*;

pub use error::*;

pub use token_bridge::context::*;

Modules

accounts

An Anchor generated module, providing a set of structs mirroring the structs deriving Accounts, where each field is a Pubkey. This is useful for specifying accounts for a client.

constants

context

error

instruction

An Anchor generated module containing the program’s set of instructions, where each method handler in the #[program] mod is associated with a struct defining the input arguments to the method. These should be used directly, when one wants to serialize Anchor instruction data, for example, when speciying instructions on a client.

message

portfolio_management

Instructions for portfolio management logic, consisting of bond management and token bridge wrapper instructions.

program

Module representing the program.

state

token_bridge

Constants

ID_CONST

Const version of ID

Statics

ID

The static program ID

Functions

check_id

Confirms that a given pubkey is equivalent to the program ID

entry

The Anchor codegen exposes a programming model where a user defines a set of methods inside of a #[program] module in a way similar to writing RPC request handlers. The macro then generates a bunch of code wrapping these user defined methods into something that can be executed on Solana.

entrypoint⚠

Safety

id

Returns the program ID

id_const

Const version of ID