Module portfolio_management::portfolio_management

Instructions for portfolio management logic, consisting of bond management and token bridge wrapper instructions.

Functions

create_bond

Creates a bond for a feed. The bond is created with a fixed sized of investors (for now), see InvestorsAccount for details.

invest_in_bond

Invests into a bond from an account. The account is placed into the investors vector inside InvestorsAccount.

redeem_bond

Redeem logic for the investors to redeem their amounts whenever they want. If the total amount amount + net_profit drops to 0 for an account then we remove it from the investors list for this bond.

redeem_wrapped_with_payload

FIXME: should be invoked within portfolio program trade

register_foreign_contract

FIXME: should be invoked within portfolio program trade

send_wrapped_with_payload

FIXME: should be invoked within portfolio program trade

token_bridge_initialize

FIXME: should be invoked within portfolio program trade

trade

TODO: still under development. Goal is to transfer using our token bridge implementation and make net profits to the investors which are then placed into the investors’ Investor net_profit fields inside the data account of investors InvestorsAccount

update_relayer_fee

FIXME: should be invoked within portfolio program trade